Avance comandas
- Se empezo frmCrearComandas y frmAgregarProductos

+++ SistemaRestaurante/SistemaRestauranteDominio/src/main/java/sistemarestaurantedominio/Producto.java
@@ -68,7 +68,7 @@ public class Producto implements Serializable {
         this.nombre = nombre;
     }
 
-    public double getPrecio() {
+    public Float getPrecio() {
         return precio;
     }
